[#796] Format DonateWidget balance, extract shared formatTokenAmount
Extract formatTokenAmount from TradingWidget to shared lib/format.ts.
Apply to DonateWidget balance and donation amount displays. Values
now show 2dp with commas for >= 1, 4dp for small, scientific for tiny.

Fixes #796
